+++ src/Adapter/PlatformServices.Desktop/Utilities/DesktopAssemblyUtility.cs
