+++ src/Adapter/PlatformServices.Desktop/Utilities/DesktopReflectionUtility.cs
