+++ src/Adapter/PlatformServices.NetCore/Utilities/NetCoreAssemblyUtility.cs
