+++ src/Adapter/PlatformServices.Shared/netstandard1.0/Services/ns10MSTestSettingsProvider.cs
