+++ src/Adapter/PlatformServices.Shared/netstandard1.0/Utilities/ns10Validate.cs
